Safer nicks, room titles now in rooms
Malicek 0.1.14 now publishes room names and topics in /rooms/:id, so we
can save one call and get more data from there.

Also, sometimes the nick list is not updated by the time the person
joins, on the backend side.  That results in the nicks dict now being
populated yet.

So we now make sure every nick has at least an empty dict with the right
keys.  If the textual ID is missing, we fall back to *@alik.cz as the
ID.  It stucks but the alternatives are worse.

Signed-off-by: Petr Šabata <contyk@redhat.com>