Reorder methods, checks, and add constants
Reorder all methods to match the order in the RFC.  This applies both
do_*() and handle_*(), as well as the check in handle().

Additionally define all the RFC numerics for the IRC enum.  Even if we
don't use them now, we might at some point.  Just thinking ahead.

Numerics are sorted by their numerical value, though, not as the RFC
lists them, which is chaotic.

Signed-off-by: Petr Šabata <contyk@redhat.com>