Send data from just one place and user standard substitutions
This is an attempt to actually use the identifier (e.g. nick)
substitutions, in case they use spaces.

We now send data from only Client.send().  reply(), error() and all the
do_*() methods now send data through this one.  It accepts the
formatting template and keyword arguments, with some predefined ones,
too.

For instance, user is encoded and defaults to self.user, and server is
encoded and defaults to config['server'].

reply() now accepts just enums, not their values, it does the
extraction, like error() did.

If one needs to pass multiple nicknames, such as when responding to
NAMES, they pass it with names, which is another special value, a list
of names to encode.

do_privmsg() decodes the message so that it's send properly, in case
such encoded names are mentioned.  This means some characters, namely
the non-breaking space, cannot be sent.  Oh well.

Signed-off-by: Petr Šabata <contyk@redhat.com>