Whispering, better CTCP, new commands
We now support '@ <nick> <message>' form of PRIVMSG for whispering.  We
try to find the nick's ID in our cache and send it as a whisper message.

We respond to bad syntax or unknown user ID with PRIVMSG.  Maybe we
should do more of those to make them obvious and bound to the actions.

We now strip SOH if it's present, not just for ACTION.  But then we
recognize only ACTION and send a PRIVMSG with error otherwise, not
sending the data.

For a given value of support, we now support DIE, RESTART, REHASH,
SUMMON and USERS.  We simply parse these and refuse to do anything.

We also support ISON, USERHOST and WALLOPS, even though WALLOPS is
ridiculous.  Might move its code to do_wallops() later and maybe use
this mechanism for various status messages.

Signed-off-by: Petr Šabata <contyk@redhat.com>