WIP: Lots of changes, mostly incomplete
This adds basic support for MODE.  Kicking, banning, locking and
reporting should be supported by the backend but needs more work here as
it's not an atomic operation on IRC.

Adds handlers for many commands, mostly dummy responses.  Notably
supports MOTDs now.

Hopefully more robust when it comes to attacks and broken clients.

Signed-off-by: Petr Šabata <contyk@redhat.com>